feat(rust): deprecate python agents, streamline compose mesh, and fix tests
