+++ _archive/python_agents/voice/resilience.py
