+++ backend/app/stt/__init__.py
@@ -1,3 +1,3 @@
-from .agent import STTAgent
+# Active STT Agent has been fully migrated to Rust.
+# Python STT implementation has been archived at _archive/python_agents/stt.
 
-__all__ = ["STTAgent"]
