+++ backend/app/voice/__init__.py
@@ -1,3 +1,3 @@
-from .agent import VoiceAgent
+# Active Voice Agent has been fully migrated to Rust.
+# Python Voice implementation has been archived at _archive/python_agents/voice.
 
-__all__ = ["VoiceAgent"]
